Refactor config types into separate files
